List of graphical parameters for Bootstrap branded plots

Sets graphical device parameters per branding guidelines. Uses _brand.yml configuration when provided explicitely via brand(), or else this package's built-in brand.

Usage

par.labs(...)

Arguments

...

Arguments passed on to graphics::par

no.readonly

logical; if TRUE and there are no other arguments, only parameters are returned which can be set by a subsequent par() call on the same device.

Value

a modified par() list

Examples

opar <- par(par.labs())
par()
#> $xlog
#> [1] FALSE
#> 
#> $ylog
#> [1] FALSE
#> 
#> $adj
#> [1] 0.5
#> 
#> $ann
#> [1] TRUE
#> 
#> $ask
#> [1] FALSE
#> 
#> $bg
#> [1] "white"
#> 
#> $bty
#> [1] "n"
#> 
#> $cex
#> [1] 1
#> 
#> $cex.axis
#> [1] 1.25
#> 
#> $cex.lab
#> [1] 1
#> 
#> $cex.main
#> [1] 1.2
#> 
#> $cex.sub
#> [1] 1
#> 
#> $cin
#> [1] 0.15 0.20
#> 
#> $col
#> [1] "#505050"
#> 
#> $col.axis
#> [1] "black"
#> 
#> $col.lab
#> [1] "black"
#> 
#> $col.main
#> [1] "black"
#> 
#> $col.sub
#> [1] "black"
#> 
#> $cra
#> [1] 10.8 14.4
#> 
#> $crt
#> [1] 0
#> 
#> $csi
#> [1] 0.2
#> 
#> $cxy
#> [1] 0.02848101 0.04109589
#> 
#> $din
#> [1] 6.666667 6.666667
#> 
#> $err
#> [1] 0
#> 
#> $family
#> [1] "Roboto Condensed"
#> 
#> $fg
#> [1] "#505050"
#> 
#> $fig
#> [1] 0 1 0 1
#> 
#> $fin
#> [1] 6.666667 6.666667
#> 
#> $font
#> [1] 1
#> 
#> $font.axis
#> [1] 1
#> 
#> $font.lab
#> [1] 1
#> 
#> $font.main
#> [1] 1
#> 
#> $font.sub
#> [1] 1
#> 
#> $lab
#> [1] 5 5 7
#> 
#> $las
#> [1] 1
#> 
#> $lend
#> [1] "round"
#> 
#> $lheight
#> [1] 1
#> 
#> $ljoin
#> [1] "round"
#> 
#> $lmitre
#> [1] 10
#> 
#> $lty
#> [1] "solid"
#> 
#> $lwd
#> [1] 1
#> 
#> $mai
#> [1] 0.8 0.4 1.0 1.0
#> 
#> $mar
#> [1] 4 2 5 5
#> 
#> $mex
#> [1] 1
#> 
#> $mfcol
#> [1] 1 1
#> 
#> $mfg
#> [1] 1 1 1 1
#> 
#> $mfrow
#> [1] 1 1
#> 
#> $mgp
#> [1] 1.0 0.5 0.0
#> 
#> $mkh
#> [1] 0.001
#> 
#> $new
#> [1] FALSE
#> 
#> $oma
#> [1] 0 0 0 0
#> 
#> $omd
#> [1] 0 1 0 1
#> 
#> $omi
#> [1] 0 0 0 0
#> 
#> $page
#> [1] TRUE
#> 
#> $pch
#> [1] 1
#> 
#> $pin
#> [1] 5.266667 4.866667
#> 
#> $plt
#> [1] 0.06 0.85 0.12 0.85
#> 
#> $ps
#> [1] 12
#> 
#> $pty
#> [1] "m"
#> 
#> $smo
#> [1] 1
#> 
#> $srt
#> [1] 0
#> 
#> $tck
#> [1] 0.025
#> 
#> $tcl
#> [1] NA
#> 
#> $usr
#> [1] 0 1 0 1
#> 
#> $xaxp
#> [1] 0 1 5
#> 
#> $xaxs
#> [1] "r"
#> 
#> $xaxt
#> [1] "s"
#> 
#> $xpd
#> [1] FALSE
#> 
#> $yaxp
#> [1] 0 1 5
#> 
#> $yaxs
#> [1] "r"
#> 
#> $yaxt
#> [1] "s"
#> 
#> $ylbias
#> [1] 0.2
#> 
# Restore device to default state
par(opar)